msm: kgsl: Check pagetable pointer before dereferencing it
Make sure that pagetable pointer is valid before dereferencing the
pointer. The pagetable pointer may be invalid if a new process is
added to the process private list and its pagetable is not yet
initialized. This can happen because a process private pointer is
added to the list and this list is guarded by kgsl driver mutex,
but the initialization of the process private pointer happens
later after the kgsl driver mutex is unlocked. So this can cause a
situation where a process private pointer is on the list but its
pagetable is not yet initialized.

CRs-Fixed: 490374
Change-Id: I2807ac557d9b640b2ca1e02f88f245cc87abcff4
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Signed-off-by: Carter Cooper <ccooper@codeaurora.org>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>