msm: kgsl: Clear dispatcher fault bit if inflight commands are 0
Clear the dispatcher fault bit if inflight commands are 0 and ensure
that the fault routine is called every time dispatcher is scheduled.
On a false hang detect there was a case when dispatcher fault routine
did not get called, this caused the dispatcher to stop issueing commands
to hardware, fix this by ensuring that fault routine is run every
time dispatcher is scheduled.

Change-Id: Ia36870e279f00630e0c79624437b0eaabb3d634c
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Signed-off-by: Carter Cooper <ccooper@codeaurora.org>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>