msm: kgsl: Re-add the irq_last functionality
irq_last is a flag to indicate kgsl should busy wait in an
attempt to NAP because the GPU is likely done with its last
batch of processing.

The setting and clearing of the irq_last variable was removed
with the dispatcher.  Although there are no visable instances
of trouble getting to NAP in current builds, re-add the functionality
until it can be fully replaced.

Change-Id: I8b4d2b72aec8948919fbeed9939684a09cb4f7f9
Signed-off-by: Lucille Sylvester <lsylvest@codeaurora.org>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>